Create CandyMachineTile component
