+++ components/index.ts
@@ -11,6 +11,7 @@ export { default as Wallet } from './Wallet'
 export { default as ExplorerLinks } from './ExplorerLinks'
 export { default as NftDetails } from './NftDetails'
 export { default as Footer } from './Footer'
+export { default as CandyMachineList } from './CandyMachineList'
 
 export * from './cards'
 export * from './forms'
