+++ pages/index.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { useWallet } from '@solana/wallet-adapter-react'
-import { CandyMachineCard, Spinner, Title } from 'components'
+import { CandyMachineList, Spinner, Title } from 'components'
 import { useRPC } from 'hooks'
 import { CANDY_MACHINE_PROGRAM_V2_ID } from 'lib/candy-machine/constants'
 import type { NextPage } from 'next'
@@ -55,12 +55,12 @@ const ManageCandyMachines: NextPage = () => {
     return (
         <>
             <Head>
-                <title>Manage Candy Machine</title>
+                <title>Candy Machines</title>
                 <meta name='description' content='Generated by create next app' />
                 <link rel='icon' href='/favicon.ico' />
             </Head>
-            <div className='container flex justify-center items-center flex-col'>
-                <Title text='Manage Candy Machine' />
+            <div className='flex-col'>
+                <h2 className='my-5'>Candy Machines · {accounts?.length}</h2>
 
                 {!error && accounts.length === 0 && <Spinner />}
                 {error && !isLoading && (
@@ -73,7 +73,7 @@ const ManageCandyMachines: NextPage = () => {
                 )}
 
                 {!isLoading && !error && accounts.length > 0 && accounts[0] !== '' && (
-                    <CandyMachineCard candyMachineAccounts={accounts} setCandyMachineAccounts={setAccounts} />
+                    <CandyMachineList candyMachineAccounts={accounts} setCandyMachineAccounts={setAccounts} />
                 )}
                 {!isLoading && !error && accounts[0] === '' && (
                     <span className='mt-5'>
